Tests/LibWeb/Text/expected/HTML/HTMLFormElement-action.txt

form.action initial value: http://www.example.com/
Final segment of form.action after setting to the empty string: HTMLFormElement-action.html
Final segment of form.action after setting to "../test.html": test.html
